+++ database/migrations/2024_03_02_222150_create_orders_table.php
@@ -11,21 +11,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('customer_id')->nullable();
-            $table->foreignId('channel_id')->nullable();
-            $table->foreignId('user_id');
-            $table->text('status', ['pending', 'processed', 'failed'])->default('pending');
-            $table->string('commentForCook')->nullable();
-            $table->string('reason')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('orders')) {
 
-            $table->foreign('customer_id')->references('id')->on('customers')->onDelete('set null');
-            // $table->foreign('payment_method_id')->references('id')->on('payment_methods')->onDelete('set null');
-            $table->foreign('channel_id')->references('id')->on('order_channels')->onDelete('set null');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('customer_id')->nullable();
+                $table->foreignId('channel_id')->nullable();
+                $table->foreignId('user_id');
+                $table->text('status', ['pending', 'processed', 'failed'])->default('pending');
+                $table->string('commentForCook')->nullable();
+                $table->string('reason')->nullable();
+                $table->timestamps();
+
+                $table->foreign('customer_id')->references('id')->on('customers')->onDelete('set null');
+                // $table->foreign('payment_method_id')->references('id')->on('payment_methods')->onDelete('set null');
+                $table->foreign('channel_id')->references('id')->on('order_channels')->onDelete('set null');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
