+++ database/migrations/2024_03_03_012836_create_order_items_table.php
@@ -11,20 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('order_items', function (Blueprint $table) {
-            $table->id();
-            $table->decimal('price');
-            $table->integer('quantity')->default(1);
-            $table->foreignId('order_id');
-            // $table->foreignId('pack_id');
-            $table->foreignId('product_id');
-            $table->unsignedInteger('package_number')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('order_items')) {
 
-            // $table->foreign('pack_id')->references('id')->on('packs')->onDelete('cascade');
-            $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-        });
+            Schema::create('order_items', function (Blueprint $table) {
+                $table->id();
+                $table->decimal('price');
+                $table->integer('quantity')->default(1);
+                $table->foreignId('order_id');
+                // $table->foreignId('pack_id');
+                $table->foreignId('product_id');
+                $table->unsignedInteger('package_number')->nullable();
+                $table->timestamps();
+
+                // $table->foreign('pack_id')->references('id')->on('packs')->onDelete('cascade');
+                $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+            });
+        }
     }
 
     /**
