+++ database/migrations/2024_03_06_144548_create_payments_table.php
@@ -11,24 +11,27 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('payments', function (Blueprint $table) {
-            $table->id();
-            // $table->decimal('amount');
-            $table->foreignId('order_id');
-            $table->foreignId('user_id');
-            // $table->foreignId('customer_id')->nullable();
-            $table->foreignId('payment_method_id')->nullable();
-            // $table->text('payment_methods')->nullable();
-            $table->decimal('paid')->default(0);
-            // $table->text('status')->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('payments')) {
 
-            $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->foreign('payment_method_id')->references('id')->on('payment_methods')->onDelete('cascade');
-            // $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
+            Schema::create('payments', function (Blueprint $table) {
+                $table->id();
+                // $table->decimal('amount');
+                $table->foreignId('order_id');
+                $table->foreignId('user_id');
+                // $table->foreignId('customer_id')->nullable();
+                $table->foreignId('payment_method_id')->nullable();
+                // $table->text('payment_methods')->nullable();
+                $table->decimal('paid')->default(0);
+                // $table->text('status')->nullable();
+                $table->timestamps();
 
-        });
+                $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->foreign('payment_method_id')->references('id')->on('payment_methods')->onDelete('cascade');
+                // $table->foreign('customer_id')->references('id')->on('customers')->onDelete('cascade');
+
+            });
+        }
     }
 
     /**
